server-blog/requirements.txt
dj-database-url==0.5.0
Django==2.0
Markdown==2.6.11
python-decouple==3.1
psycopg2==2.7.1
djangorestframework
django-cors-headers
